+++ CHANGES.rst
@@ -4,4 +4,5 @@ Changelog
 1.0a1 (unreleased)
 -------------------
 
-- Initial release
+- Added fields parameter in Mailer override expressions
+  [gborelli]